fix: correct CI build/lint failures from the tsdown migration
Fixes several real bugs the CI run surfaced on top of the initial
migration commit:

- `external()`'s new self-reference bypass ran before the check for
  explicitly-declared manual externals, incorrectly overriding packages
  (like `@ember-data/debug`) that deliberately list a self-referencing
  subpath as external. Check declared externals first.
- `core`'s `eslint.config.mjs`/`typedoc.config.mjs` still imported from
  the now-deleted `vite.config.mjs`.
- `core`'s curated entry-point list was missing several subpaths
  (`signals/-private`, `reactive/-private/schema`,
  `reactive/-private/fields/managed-array`,
  `store/-private/network/request-cache`) that other packages deep-import
  but that never needed to be "official" entries under the old
  isolated-declarations approach (which emitted a declaration for every
  source file regardless). Under the new rolled-up declarations, an
  uncovered subpath silently loses its types instead.
- `store/deprecated/store.ts`'s `declare module` cross-file augmentation
  (adding `findRecord`/`saveRecord`/`modelFor`/etc. to `Store`) doesn't
  survive rolldown-plugin-dts's entry-based bundling: `Store` gets hoisted
  into a shared, hash-named chunk whenever 2+ entries reference it (always),
  and a cross-file augmentation can't target an unstable chunk name -- it
  only ever attached for consumers importing through the one exact path
  the augmentation happened to specify, typing these methods as `any` for
  every real consumer otherwise. Moved the interface (types only, runtime
  implementation unchanged) into `store-service.ts` as a second, clearly-
  commented `Store` interface -- same-module interface merging has no such
  problem, since it resolves before any bundling happens.
- The four `@warp-drive/build-config` babel macro transforms
  (assert/deprecations/features/logging) crashed on `import * as X`
  specifiers -- rolldown emits this as a defensive fallback alongside a
  literal `export * from '...'` re-export, and none of the transforms
  guarded against a specifier that isn't a plain named import.
- Removed now-stale `eslint-disable` comments in `legacy` and `specs`
  that were suppressing the `any`-typing fallout from the above.

Verified with a full clean `pnpm install` (35/35 build tasks), full
`turbo run lint` (82/82), and the `core` test suite (187/189, 0 failures,
2 pre-existing skips).
