+++ specs/src/request-component.spec.ts
@@ -620,7 +620,6 @@ export const RequestSpec: SuiteBuilder<LocalTestContext, RequestSpecSignature> =
     assert.dom().hasText('Chris ThoburnCount: 3');
 
     // ensure background request has completed
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-call
     await store._getAllPending();
     await this.h.rerender();
 
@@ -1157,7 +1156,6 @@ export const RequestSpec: SuiteBuilder<LocalTestContext, RequestSpecSignature> =
     await new Promise((resolve) => setTimeout(resolve, DEBUG ? 1 : 100));
 
     if (PRODUCTION) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-call
       await store._getAllPending();
       await this.h.rerender();
     } else {
@@ -1320,7 +1318,6 @@ export const RequestSpec: SuiteBuilder<LocalTestContext, RequestSpecSignature> =
     assert.equal(record!.name, 'Chris Thoburn');
     assert.verifySteps(['set-request', 'Chris Thoburn']);
 
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-call
     await store._getAllPending();
     await this.h.rerender();
 
@@ -1398,7 +1395,6 @@ export const RequestSpec: SuiteBuilder<LocalTestContext, RequestSpecSignature> =
     assert.equal(record!.name, 'Chris Thoburn');
     assert.verifySteps(['set-request', 'loading']);
 
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-call
     await store._getAllPending();
     await this.h.rerender();
