+++ tools/internal-config/rollup/external.js
@@ -66,6 +66,16 @@ export function external(manual = []) {
 
   // console.log({ externals: result });
   return function (id) {
+    // An explicit (manual) external, or a declared dependency/peerDependency,
+    // always wins -- even if it also happens to be a self-referencing
+    // subpath of this very package (e.g. `@ember-data/debug` deliberately
+    // lists `@ember-data/debug/data-adapter` as external so it stays a bare
+    // specifier for the consuming app to resolve, rather than being routed
+    // to an internal entry). Check this before the self-reference bypass below.
+    if (all.has(id)) {
+      return true;
+    }
+
     // A package importing its own name (e.g. `@warp-drive/core` source
     // importing `@warp-drive/core/build-config/env`) is a self-reference, not
     // an external dependency. Under Vite/Rollup this branch is never reached
@@ -76,10 +86,6 @@ export function external(manual = []) {
     // resolution proceed so a self-reference-aware resolver plugin can resolve it.
     if (id === pkg.name || id.startsWith(pkg.name + '/')) {
       return false;
-    }
-
-    if (all.has(id)) {
-      return true;
     }
 
     for (const dep of deps) {
